fix: 将kline_quality.py中的"vol"数据类型从"int64"更改为"float"，以确保数据一致性

+++ czsc/utils/kline_quality.py
@@ -39,7 +39,7 @@ def check_data_types(df):
         "close": "float",
         "high": "float",
         "low": "float",
-        "vol": "int64",
+        "vol": "float",
         "amount": "float",
     }
     type_mismatches = {}
